refactor(settings): extract disable/delete account into useAccountActions
Move the two destructive account paths (confirm -> re-auth -> service ->
redirect, delete adds a typed-DELETE prompt) verbatim into
_hooks/use-account-actions.ts, isolated from the form-bound saves since
they're the highest-stakes flows. Owns its own router + Settings i18n;
takes requestVerification + message helpers. Behavior-identical;
page.tsx 1458->1430 LOC.
